Clean up formatting sys info for test report
